+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-## [0.1.3-preview.3] - 2019-01-29
+## [0.1.3-preview.4] - 2019-01-29
 - Add typing controller.
 - Add event delay.
 - Add event end.
+- Add Choice system.
+- Load scene event now support load by name.
 
 ## [0.1.2-preview.6] - 2019-01-28
 - Update support DSC UI to V0.0.2-preview.1